refactor: streak update 로직 분리
