+++ src/main/java/oncog/cogroom/domain/streak/scheduler/StreakSchduler.java
@@ -3,6 +3,7 @@ package oncog.cogroom.domain.streak.scheduler;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import oncog.cogroom.domain.streak.service.StreakService;
+import oncog.cogroom.domain.streak.service.StreakUpdateService;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
@@ -11,13 +12,13 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class StreakSchduler {
 
-    private final StreakService streakService;
+    private final StreakUpdateService streakUpdateService;
 
     @Scheduled(cron = "0 0 0 * * *") // 자정마다
 //    @Scheduled(cron = "0 */2 * * * *") // 2분마다, 테스트용
     public void updateStreakAtMidnight() {
         log.info("스트릭 기록 초기화 시작");
-        streakService.updateAllMemberStreaks();
+        streakUpdateService.updateAllMemberStreaks();
         log.info("스트릭 기록 초기화 완료");
     }
 }
