+++ src/main/java/oncog/cogroom/domain/streak/service/StreakService.java
@@ -28,26 +28,6 @@ public class StreakService extends BaseService {
     private final StreakLogRepository streakLogRepository;
     private final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
 
-    @Transactional
-    public void updateAllMemberStreaks() {
-
-        LocalDateTime startOfYesterday = getStartOfYesterday();
-        LocalDateTime endOfYesterday = getEndOfYesterday();
-
-        List<Streak> streaks = streakRepository.findAll();
-
-        streaks.forEach(streak -> { // 추후 배치 적용 필요
-            Member member = streak.getMember();
-
-           boolean hasYesterdayLog = hasLogForYesterday(member, startOfYesterday, endOfYesterday);
-
-           // 해당 멤버가 전날에 작성한 log 정보가 없는 경우 누적 스트릭 일수를 0으로 초기화
-           if (!hasYesterdayLog && streak.getTotalDays() > 0) { // 기존에 0이 아닐 때만 0으로 초기화 (불필요한 업데이트 방지)
-               streak.resetTotalDays();
-           }
-        });
-    }
-
     public StreakCalendarResponseDTO getStreakDates() {
         Member member = getMember();
 
@@ -67,18 +47,6 @@ public class StreakService extends BaseService {
                 .streakDays(streakDays)
                 .streakDateList(streakDates)
                 .build();
-    }
-
-    private boolean hasLogForYesterday(Member member, LocalDateTime start, LocalDateTime end) {
-        return streakLogRepository.existsByMemberAndCreatedAtBetween(member, start, end);
-    }
-
-    private LocalDateTime getStartOfYesterday() {
-        return LocalDateTime.now().minusDays(1).toLocalDate().atStartOfDay();
-    }
-
-    private LocalDateTime getEndOfYesterday() {
-        return getStartOfYesterday().plusDays(1).minusNanos(1);
     }
 
     private LocalDateTime getStartOfCalendarMonth() {
@@ -109,6 +77,7 @@ public class StreakService extends BaseService {
         streakLogRepository.save(StreakLog.builder().member(member).streak(streak).build());
     }
 
+    // 스트릭 연속 일자 조회 (totalDays)
     public int getStreakDays(Member member) {
         return streakRepository.findByMember(member)
                 .map(Streak::getTotalDays)
